+++ erpnext/hr/doctype/salary_slip_loan/salary_slip_loan.json
@@ -7,7 +7,7 @@
  "field_order": [
   "loan",
   "loan_type",
-  "column_break_0b6yy",
+  "column_break_4",
   "loan_account",
   "loan_repayment_date",
   "section_break_qtz1f",
@@ -46,6 +46,10 @@
    "label": "Interest Income Account",
    "options": "Account",
    "read_only": 1
+  },
+  {
+   "fieldname": "column_break_4",
+   "fieldtype": "Column Break"
   },
   {
    "fieldname": "principal_amount",
@@ -106,10 +110,6 @@
    "read_only": 1
   },
   {
-   "fieldname": "column_break_0b6yy",
-   "fieldtype": "Column Break"
-  },
-  {
    "fieldname": "section_break_qtz1f",
    "fieldtype": "Section Break"
   },
@@ -139,7 +139,7 @@
  ],
  "istable": 1,
  "links": [],
- "modified": "2023-11-16 14:45:14.093813",
+ "modified": "2023-11-16 14:45:15.093813",
  "modified_by": "Administrator",
  "module": "HR",
  "name": "Salary Slip Loan",
